category selection in restaurant view now works

+++ src/at/fhooe/im620/restaurantfinder/dao/GenericDAO.java
@@ -5,7 +5,7 @@ import java.util.List;
 public interface GenericDAO<T> {
 	
 	public void saveOrUpdateEntity(T entity);
-	public T getEntityById(int id);
+	public T getEntityById(Long id);
 	public List<T> getAllEntities();
 	public void deleteEntity(T entity);
